scripts -> cli
