+++ cli/_utils/fs.ts
@@ -1,0 +1,4 @@
+import _fs from "fs";
+
+export { _fs };
+export const fs = _fs.promises;
