+++ cli/_utils/projectPath.ts
@@ -1,0 +1,3 @@
+import path from "path";
+
+export const projectPath = path.resolve(__dirname, "..", "..");
