+++ cli/_utils/registerBabel.js
@@ -1,0 +1,13 @@
+require("@babel/register")({
+	presets: [
+		[
+			"@babel/preset-env",
+			{
+				targets: { node: 12 },
+			},
+		],
+		"@babel/preset-typescript",
+	],
+	plugins: ["@babel/plugin-transform-runtime"],
+	extensions: [".js", ".ts"],
+});
