+++ cli/_utils/spawnStdio.ts
@@ -1,0 +1,1 @@
+export const spawnStdio = [process.stdin, process.stdout, process.stderr];
