+++ cli/index.js
@@ -1,0 +1,2 @@
+require("./_utils/registerBabel");
+require("./_index");
