--- nodemon.json
@@ -1,6 +1,0 @@
-{
-	"ignore": [".git", "node_modules"],
-	"watch": ["src"],
-	"exec": "npm run start:dev",
-	"ext": "ts"
-}
