+++ ormconfig.js
@@ -1,4 +1,4 @@
 require("dotenv").config();
-require("./scripts/babel");
+require("./cli/_utils/registerBabel");
 
 module.exports = require("./src/appConfig").appConfig.dbConnectionOptions;
