+++ package.json
@@ -1,15 +1,14 @@
 {
 	"name": "openfing-server",
-	"version": "1.1.2",
+	"version": "1.1.3",
 	"description": "OpenFING Server",
 	"scripts": {
-		"start:dev": "node -r ./scripts/babel ./src",
-		"start:dev:watch": "nodemon",
-		"generate-files": "node -r ./scripts/babel ./scripts/generate-files",
-		"build": "babel src --out-dir dist --copy-files --extensions '.ts,.js'",
-		"deploy": "node -r ./scripts/babel ./scripts/deploy",
-		"postinstall": "patch-package",
-		"typeorm": "node --require ts-node/register ./node_modules/typeorm/cli.js"
+		"start-dev": "node cli start-dev",
+		"start-dev:watch": "node cli start-dev -w",
+		"generate-files": "node cli generate-files",
+		"build": "node cli build",
+		"deploy": "node cli deploy",
+		"postinstall": "patch-package"
 	},
 	"repository": {
 		"type": "git",
@@ -41,10 +40,12 @@
 		"@types/lodash": "4.14.161",
 		"@types/moment-timezone": "0.5.30",
 		"@types/node": "12.12.55",
+		"@types/nodemon": "1.19.0",
 		"@types/pg": "7.14.4",
 		"@types/pg-copy-streams": "1.2.1",
 		"@types/url-join": "4.0.0",
 		"@types/uuid": "8.3.0",
+		"@types/yargs": "15.0.5",
 		"@types/yup": "0.29.6",
 		"@typescript-eslint/eslint-plugin": "4.0.1",
 		"@typescript-eslint/parser": "4.0.1",
@@ -52,13 +53,13 @@
 		"eslint-config-prettier": "6.11.0",
 		"eslint-plugin-prettier": "3.1.4",
 		"eslint-plugin-simple-import-sort": "5.0.3",
-		"mobx": "5.15.6",
 		"node": "12.18.3",
 		"nodemon": "1.19.4",
 		"patch-package": "6.2.2",
 		"prettier": "2.1.1",
 		"ts-node": "9.0.0",
-		"typescript": "4.0.2"
+		"typescript": "4.0.2",
+		"yargs": "16.0.3"
 	},
 	"dependencies": {
 		"@babel/runtime": "7.11.2",
@@ -82,6 +83,7 @@
 		"moment-timezone": "0.5.31",
 		"pg": "8.3.3",
 		"pg-copy-streams": "5.1.1",
+		"promisify-child-process": "4.1.1",
 		"reflect-metadata": "0.1.13",
 		"ssh2-promise": "0.1.7",
 		"typeorm": "0.2.25",
