--- scripts/babel.js
@@ -1,2 +1,0 @@
-require("@babel/register")({ extensions: [".js", ".ts"] });
-require("../src/_utils/configEnv");
