--- scripts/deploy/index.ts
@@ -1,1 +1,0 @@
-import "./deploy";
